Update TrackTodayExercises.py
fix exercise_id format int

+++ WORKOUT_MEAL_WEEKLY_PLAN/pages/TrackTodayExercises.py
@@ -93,7 +93,8 @@ def display_track_exercises_page(st):
         tracking_response = supabase.table("exercise_tracking").select("*").eq("user_id", st.session_state.user['user_id']).gte("completed_date", start_str).lte("completed_date", end_str).execute()
         completed_df = pd.DataFrame(tracking_response.data)
 
-            
+        #fix exercise_id format
+        completed_df['exercise_id'] = completed_df['exercise_id'].astype(int)
 
         if completed_df.empty:
             return []  # none completed, return all
